+++ CHANGELOG.md
@@ -51,6 +51,8 @@ myColorField: {
 }
 ```
 * Restored Vue dependency to using semantic versioning now that Vue 2.6.14 has been released with a fix for the bug that required us to pin 2.6.12.
+* `min` and `max` work properly for both string-like and number-like fields.
+* Negative numbers, leading minus and plus signs, and trailing periods are accepted in the right ways by appropriate field types.
 
 ## 3.0.0-beta.3.1 - 2021-06-07